Moving TreeStorage to separate directory
